+++ client/src/app/views/chapters/chapters-list/chapters-list.component.ts
@@ -3,7 +3,7 @@ import { Chapter } from '../../../models/chapter.model';
 import { ChaptersService } from '../../../services/chapters.service';
 import { EditChapterComponent } from '../edit-chapter/edit-chapter.component';
 import { MatDialog } from "@angular/material/dialog";
-
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-chapters-list',
@@ -29,10 +29,29 @@ export class ChaptersListComponent implements OnInit {
   }
 
   deleteChapter(id): void {
-    this.chaptersService.deleteChapter(id).subscribe((res) => {
-      console.log(res);
-      this.getChapters();
+
+    Swal.fire({
+      title: 'Vous êtes sûr?',
+      text: "Les données supprimées seront perdus",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Annuler',
+      confirmButtonText: 'Oui, supprimer!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Supprimé!',
+          'Suppression terminée avec succès',
+          'success'
+        )
+        this.chaptersService.deleteChapter(id).subscribe((res) => {
+          this.getChapters();
+        })
+      }
     })
+
   }
 
   openEditChapter(chapter: Chapter): void {
